IFSintegrals.barycentre_rule — Function

x,w = barycentre_rule(Γ::Union{InvariantMeasure,SubInvariantMeasure},h::Real)

returns a vector of N weights wⱼ>0 and nodes xⱼ ∈ Rⁿ, for approximation of integrals defined on an IFS Γ⊂Rⁿ.

x,y,w = barycentre_rule(Γ₁::Union{InvariantMeasure,SubInvariantMeasure},Γ₂::Union{InvariantMeasure,SubInvariantMeasure},h::Real)

returns N weights wⱼ>0 and nodes x,y ∈ Rⁿ, for approximation of double integrals over Γ₁,Γ₂⊂Rⁿ. Uses Barycentre rule quadrature, the fractal Γ will be subdivided until each subcomponent has a diameter of less than h.

IFSintegrals.gauss_quad — Function

x,w = gauss_quad(Γ::SelfSimilarFractal{V,M}, N::Int64) where {V<:Real, M<:Real}

Returns N Gaussian weights w ∈ Rᴺ and nodes x ∈ Rᴺˣᴺ. Here Γ must be an SelfSimilarFractal in one spatial dimension. N is the order of the Gauss rule, i.e. number of weights and nodes.

IFSintegrals.chaos_quad — Function

chaos_quad(Γ::SelfSimilarFractal{V,M},N::Int64;x₀=Γ.barycentre:::AbstractVector) where {V<:AbstractVector, M<:Union{Real,AbstractMatrix}}

Returns a vector of N weights w>0 and nodes x ∈ Rⁿ, for approximation of integrals defined on an IFS Γ⊂Rⁿ. Using Chaos game quadrature. Optional third input is the initial guess, which is taken as barycentre by default.