An iterated function system is a set of $M$ affine contraction maps, or similarities.

Each similarity is of the form:

\[s_m(x)=r_mA_mx + \delta_m,\quad x\in\mathbb{R}^n,\]

where $r_m\in(0,1)$ is the contraction factor, $A_m\in\R^{n\times n}$ is a rotation matrix, and $\delta\in\R^{n}$ is a translation vector.

IFSintegrals.Similarity — Type

struct Similarity{V<:Union{Real,AbstractVector}, M<:Union{Real,AbstractMatrix}}
    r::Float64 # contraction
    δ::V # translation
    A::M # rotation
    rA::M # contraction * rotation
end

Constructs a similarity map. The third input (rotation matrix) is optional, and the fourth is created automatically.

Can be treated as a function, for example

    julia> s = Similarity(1/2,0) # creates contraction s(x)=x/2+0
    julia> s(π)
    1.5707963267948966

The iterated function system may be interpreted as a set of these maps, $\{s_m\}_{m=1}^M$. The attractor $\Gamma$ is the unique non-empty compact set which satisfies $ \Gamma = \bigcup{m=1}^M sm(\Gamma). $

The attractor $\Gamma$ will be the support of our measure. There are other properties, such as diameter and Hausdorff dimension, which describe this type.

Missing docstring.

Missing docstring for SelfSimilarMeasure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SelfSimilarSubMeasure. Check Documenter's build log for details.